requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
PyMuPDF
aiofiles